+++ packages/@headlessng-primitives/CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this package will be documented in this file.
 ## [Unreleased]
 
 ### Added
---
+- `AutoFocus` directive added
+
+### Changed
+
+-
 
 ### Fixed
---
+
+-